requirements.txt
Jinja2==3.1.5
lxml==5.3.0
MarkupSafe==3.0.2
python-docx==1.1.2
PyYAML==6.0.2
typing_extensions==4.12.2
